64ビット対応と画像インポートのバグ修正
64ビット版Windowsに対応した。また、画像インポートに存在した大文字拡張子のファイル読み込みのバグを修正した。

+++ ComicFileHandler/CRegisterExtension.cpp
@@ -59,7 +59,7 @@ HRESULT CRegisterExtension::RegisterExtensionWithProgID(const wstring& wstrFileE
 
 HRESULT CRegisterExtension::SetRegValue(HKEY hkey, const wstring& wstrKeyName, const wstring& wstrValueName, const wstring& wstrValue)
 {
-	TEST(HRESULT_FROM_WIN32(static_cast<unsigned long>(RegSetKeyValueW(hkey, wstrKeyName.c_str(), wstrValueName.c_str(), REG_SZ, wstrValue.c_str(), (wstrValue.length() + 1) * sizeof(WCHAR)))));
+	TEST(HRESULT_FROM_WIN32(static_cast<unsigned long>(RegSetKeyValueW(hkey, wstrKeyName.c_str(), wstrValueName.c_str(), REG_SZ, wstrValue.c_str(), (static_cast<DWORD>((wstrValue.length() + 1) * sizeof(WCHAR)))))));
 	UpdateAssocChanged(hkey, wstrKeyName);
 	return S_OK;
 }
